requirements3.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements3.txt requirements3.in
#

appdirs==1.4.3            # via setuptools
click==6.7                # via pip-tools
couchdb==1.1
first==2.0.1              # via pip-tools
invoke==0.12.2
jinja2==2.9.6
MarkupSafe==1.0           # via jinja2
nose==1.3.7
packaging==16.8           # via setuptools
pip-tools==1.9.0
pkginfo==1.4.1            # via twine
pyparsing==2.2.0          # via packaging
pyyaml==3.12
requests-toolbelt==0.8.0  # via twine
requests==2.14.2
six==1.10.0               # via packaging, pip-tools, setuptools
tqdm==4.11.2              # via twine
twine==1.9.0

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via twine
